Simplify the code by distributing it to the right folders
