+++ src/charm.py
@@ -16,450 +16,37 @@ the user.
 """
 
 import logging
-import os
 import subprocess
-from functools import cached_property
-from math import ceil
-from typing import Any
 
 import charms.operator_libs_linux.v0.apt as apt
 import ops
-from charms.data_platform_libs.v0.data_interfaces import KafkaRequires
-from charms.kafka.v0.client import KafkaClient, NewTopic
-from ops.charm import CharmBase
 from ops.framework import EventBase
-from ops.model import BlockedStatus, Unit
 from overrides import override
-from tenacity import Retrying, stop_after_attempt, wait_fixed
 
 from benchmark.base_charm import DPBenchmarkCharmBase
-from benchmark.core.models import (
-    DatabaseState,
-    PeerState,
-)
 from benchmark.core.pebble_workload_base import DPBenchmarkPebbleWorkloadBase
-from benchmark.core.structured_config import BenchmarkCharmConfig
 from benchmark.core.systemd_workload_base import DPBenchmarkSystemdWorkloadBase
 from benchmark.core.workload_base import WorkloadBase
-from benchmark.events.actions import ActionsHandler
-from benchmark.events.db import DatabaseRelationHandler
-from benchmark.events.peer import PeerRelationHandler
 from benchmark.literals import (
     PEER_RELATION,
     DPBenchmarkLifecycleState,
-    DPBenchmarkLifecycleTransition,
 )
-from benchmark.managers.config import ConfigManager
-from benchmark.managers.lifecycle import LifecycleManager
-from core.models import KafkaBenchmarkCharmConfig, KafkaDatabaseState, WorkloadTypeParameters
+from core.models import KafkaBenchmarkCharmConfig
+from events.actions import KafkaBenchmarkActionsHandler
+from events.kafka import KafkaDatabaseRelationHandler
+from events.peers import KafkaPeersRelationHandler
 
 # TODO: This line must go away once Kafka starts sharing its certificates via client relation
-from events.tls import JavaTlsHandler, JavaTlsStoreManager
+from events.tls import JavaTlsHandler
 from literals import (
     CLIENT_RELATION_NAME,
-    INITIAL_PORT,
     JAVA_VERSION,
-    PORT_JUMP,
-    TEN_YEARS_IN_MINUTES,
-    TOPIC_NAME,
-    WORKER_PARAMS_YAML_FILE,
 )
+from managers.config import KafkaConfigManager
+from managers.lifecycle import KafkaLifecycleManager
 
 # Log messages can be retrieved using juju debug-log
 logger = logging.getLogger(__name__)
-
-
-class KafkaDatabaseRelationHandler(DatabaseRelationHandler):
-    """Listens to all the DB-related events and react to them.
-
-    This class will provide the charm with the necessary data to connect to the DB as
-    well as the current relation status.
-    """
-
-    DATABASE_KEY = "topic"
-
-    def __init__(
-        self,
-        charm: CharmBase,
-        relation_name: str,
-    ):
-        super().__init__(charm, relation_name)
-        self.consumer_prefix = f"{charm.model.name}-{charm.app.name}-benchmark-consumer"
-        # We can have only one reference to a Requires object.
-        # Hence, we will store it here:
-        self._internal_client = KafkaRequires(
-            self.charm,
-            self.relation_name,
-            TOPIC_NAME,
-            extra_user_roles="admin",
-            consumer_group_prefix=self.consumer_prefix,
-        )
-
-    @property
-    @override
-    def state(self) -> KafkaDatabaseState:
-        """Returns the state of the database."""
-        if not (self.relation and self.client):
-            logger.error("Relation data not found")
-            # We may have an error if the relation is gone but self.relation.id still exists
-            # or if the relation is not found in the fetch_relation_data yet
-            return KafkaDatabaseState(
-                self.charm.app,
-                None,
-            )
-
-        return KafkaDatabaseState(
-            self.charm.app,
-            self.relation,
-            data=self.client.fetch_relation_data().get(self.relation.id, {}),
-        )
-
-    @property
-    @override
-    def client(self) -> Any:
-        """Returns the data_interfaces client corresponding to the database."""
-        return self._internal_client
-
-    def bootstrap_servers(self) -> list[str] | None:
-        """Return the bootstrap servers."""
-        if not self.state or not (model := self.state.model()):
-            return None
-        return model.hosts
-
-    def tls(self) -> tuple[str | None, str | None]:
-        """Return the TLS certificates."""
-        if not self.state.tls_ca:
-            return self.state.tls, None
-        return self.state.tls, self.state.tls_ca
-
-
-class KafkaPeersRelationHandler(PeerRelationHandler):
-    """Listens to all the peer-related events and react to them."""
-
-    @override
-    def peers(self) -> list[str]:
-        """Return the peers."""
-        if not self.relation:
-            return []
-        return [
-            f"{self.relation.data[u]['ingress-address']}:{port}"
-            for u in list(self.units()) + [self.this_unit()]
-            for port in range(
-                INITIAL_PORT,
-                INITIAL_PORT + PORT_JUMP * self.charm.config.parallel_processes,
-                PORT_JUMP,
-            )
-        ]
-
-
-class KafkaConfigManager(ConfigManager):
-    """The config manager class."""
-
-    def __init__(
-        self,
-        workload: WorkloadBase,
-        database_state: DatabaseState,
-        peer_state: PeerState,
-        java_tls: JavaTlsStoreManager,
-        peers: list[str],
-        config: BenchmarkCharmConfig,
-        is_leader: bool,
-        labels: str = "",
-    ):
-        super().__init__(workload, database_state, peer_state, peers, config, labels, is_leader)
-        self.worker_params_template_file = "kafka_worker_params_template.j2"
-        self.java_tls = java_tls
-        self.is_leader = is_leader
-        self.systemd_service_template_file = "kafka_benchmark.service.j2"
-
-    def _service_args(
-        self, args: dict[str, Any], transition: DPBenchmarkLifecycleTransition
-    ) -> dict[str, Any]:
-        return args | {
-            "charm_root": self.workload.paths.charm_dir,
-            "command": transition.value,
-            "is_coordinator": "" if not self.is_leader else "True",
-        }
-
-    @override
-    def _render_service(
-        self,
-        transition: DPBenchmarkLifecycleTransition,
-        dst_path: str | None = None,
-    ) -> str | None:
-        """Render the workload parameters."""
-        if not (options := self.get_execution_options()):
-            return None
-        values = self._service_args(options.dict(), transition)
-        return self._render(
-            values=values,
-            template_file=self.systemd_service_template_file,
-            template_content=None,
-            dst_filepath=dst_path,
-        )
-
-    @override
-    def _check(
-        self,
-        transition: DPBenchmarkLifecycleTransition,
-    ) -> bool:
-        if not (
-            os.path.exists(self.workload.paths.service)
-            and os.path.exists(self.workload.paths.workload_params)
-            and (values := self.get_execution_options())
-        ):
-            return False
-        values = self._service_args(values.dict(), transition)
-        compare_svc = "\n".join(self.workload.read(self.workload.paths.service)) == self._render(
-            values=values,
-            template_file=self.systemd_service_template_file,
-            template_content=None,
-            dst_filepath=None,
-        )
-        compare_params = "\n".join(
-            self.workload.read(self.workload.paths.workload_params)
-        ) == self._render(
-            values=self.get_workload_params(),
-            template_file=None,
-            template_content=self.workload.workload_params_template,
-            dst_filepath=None,
-        )
-        return compare_svc and compare_params
-
-    def get_worker_params(self) -> dict[str, Any]:
-        """Return the workload parameters."""
-        # Generate the truststore, if applicable
-        self.java_tls.set_truststore()
-        if not (db := self.database_state.model()):
-            return {}
-        num_brokers = len(db.hosts) if db.hosts else 0
-        return {
-            "total_num_replicas": min(num_brokers, 3),
-            # We cannot have quotes nor brackets in this string.
-            # Therefore, we render the entire line instead
-            "list_of_brokers_bootstrap": "bootstrap.servers={}".format(
-                ",".join(db.hosts) if db.hosts else ""
-            ),
-            "username": db.username,
-            "password": db.password,
-            "threads": self.config.threads if self.config.threads > 0 else 1,
-            "truststore_path": self.java_tls.java_paths.truststore,
-            "truststore_pwd": self.java_tls.truststore_pwd,
-            "ssl_client_auth": "none",
-        }
-
-    def _render_worker_params(
-        self,
-        dst_path: str | None = None,
-    ) -> str | None:
-        """Render the worker parameters."""
-        return self._render(
-            values=self.get_worker_params(),
-            template_file=self.worker_params_template_file,
-            template_content=None,
-            dst_filepath=dst_path,
-        )
-
-    @override
-    def get_workload_params(self) -> dict[str, Any]:
-        """Return the worker parameters."""
-        workload = WorkloadTypeParameters[self.config.workload_name]
-        clients = ceil((int(self.config.parallel_processes) * len(self.peers)) / 2)
-        partitions_per_topic = self.config.threads * clients
-
-        return {
-            "partitionsPerTopic": partitions_per_topic,
-            "clients": clients,
-            "duration": int(self.config.duration / 60)
-            if self.config.duration > 0
-            else TEN_YEARS_IN_MINUTES,
-            "charm_root": self.workload.paths.charm_dir,
-            "producerRate": workload.producer_rate,
-            "messageSize": workload.message_size,
-        }
-
-    @override
-    def _render_params(
-        self,
-        dst_path: str,
-    ) -> str | None:
-        """Render the workload parameters.
-
-        Overloaded as we need more than one file to be rendered.
-
-        This extra file will rendered in the same folder as `dst_path`, but with a different name.
-        """
-        super()._render_params(dst_path)
-        self._render_worker_params(
-            os.path.join(
-                os.path.dirname(os.path.abspath(dst_path)),
-                WORKER_PARAMS_YAML_FILE,
-            )
-        )
-
-    @override
-    def prepare(self) -> bool:
-        """Prepare the benchmark service."""
-        super().prepare()
-        try:
-            for attempt in Retrying(
-                stop=stop_after_attempt(4), wait=wait_fixed(wait=20), reraise=True
-            ):
-                with attempt:
-                    if (model := self.database_state.model()) and not self.is_prepared():
-                        topic = NewTopic(
-                            name=model.db_name,
-                            num_partitions=self.config.parallel_processes * (len(self.peers) + 1),
-                            replication_factor=self.client.replication_factor,
-                        )
-                        self.client.create_topic(topic)
-        except Exception as e:
-            logger.debug(f"Error creating topic: {e}")
-
-        # We may fail to create the topic, as the relation has been recently stablished
-        return self.is_prepared()
-
-    @override
-    def is_prepared(self) -> bool:
-        """Checks if the benchmark service has passed its "prepare" status."""
-        try:
-            if model := self.database_state.model():
-                return model.db_name in self.client._admin_client.list_topics()
-        except Exception as e:
-            logger.info(f"Error describing topic: {e}")
-        return False
-
-    @override
-    def clean(self) -> bool:
-        """Clean the benchmark service."""
-        if not self.is_leader:
-            return super().clean()
-        # Only applicable for the leader unit
-        try:
-            if model := self.database_state.model():
-                self.client.delete_topics([model.db_name])
-            self.workload.remove(self.workload.paths.service)
-            self.workload.reload()
-
-        except Exception as e:
-            logger.info(f"Error deleting topic: {e}")
-        return self.is_cleaned()
-
-    @override
-    def is_cleaned(self) -> bool:
-        """Checks if the benchmark service has passed its "prepare" status."""
-        if not self.is_leader:
-            # For a non leader unit, it cannot access kafka as the fetch_relation_data does not return credentials
-            # Therefore, for these units, we use the standard check
-            return super().is_cleaned()
-
-        try:
-            if not (model := self.database_state.model()):
-                return False
-            return model.db_name not in self.client._admin_client.list_topics()
-        except Exception as e:
-            logger.info(f"Error describing topic: {e}")
-            return False
-
-    @cached_property
-    def client(self) -> KafkaClient:
-        """Return the Kafka client."""
-        has_tls_ca = False
-        if os.path.exists(self.java_tls.java_paths.ca):
-            has_tls_ca = True
-
-        if not (state := self.database_state.model()):
-            return KafkaClient(
-                servers=[],
-                username=None,
-                password=None,
-                security_protocol="SASL_PLAINTEXT",
-                replication_factor=1,
-            )
-
-        host_count = len(state.hosts) if state.hosts else 1
-        return KafkaClient(
-            servers=state.hosts or [],
-            username=state.username,
-            password=state.password,
-            security_protocol="SASL_SSL" if has_tls_ca else "SASL_PLAINTEXT",
-            cafile_path=self.java_tls.java_paths.ca,
-            certfile_path=None,
-            replication_factor=host_count - 1,
-        )
-
-
-class KafkaBenchmarkActionsHandler(ActionsHandler):
-    """Handle the actions for the benchmark charm."""
-
-    def __init__(self, charm: DPBenchmarkCharmBase):
-        """Initialize the class."""
-        super().__init__(charm)
-        self.config: BenchmarkCharmConfig = charm.config
-
-    @override
-    def _preflight_checks(self) -> bool:
-        """Check if we have the necessary relations.
-
-        In kafka case, we need the client relation to be able to connect to the database.
-        """
-        if (
-            int(self.config.parallel_processes) * len(self.charm.peers.all_unit_states().keys())
-            < 2
-        ):
-            logger.error("The number of parallel processes must be greater than 1.")
-            self.unit.status = BlockedStatus(
-                "The number of parallel processes must be greater than 1."
-            )
-            return False
-        return super()._preflight_checks()
-
-
-class KafkaLifecycleManager(LifecycleManager):
-    """The lifecycle manager class.
-
-    There is one extra case we must consider: before moving to running, we need to ensure
-    all the peers are in the same state are already running.
-    """
-
-    def __init__(
-        self,
-        peers: dict[Unit, PeerState],
-        this_unit: Unit,
-        config_manager: ConfigManager,
-        is_leader: bool = False,
-    ):
-        super().__init__(peers, this_unit, config_manager, is_leader)
-        self.is_leader = is_leader
-
-    @override
-    def _peers_state(self) -> DPBenchmarkLifecycleState | None:
-        next_state = super()._peers_state()
-
-        if next_state == DPBenchmarkLifecycleState.STOPPED and not self.is_leader:
-            # If not leader, then we can only stop if the leader has issued a stop directive
-            if self.peers[self.this_unit].stop_directive:
-                return DPBenchmarkLifecycleState.STOPPED
-            return None
-
-        # Now, there is a special rule: if we are the leader unit, we can only move to RUNNING
-        # if all the peers are already running or if peer count is zero
-        if not self.is_leader:
-            return next_state
-
-        # Check the special case for running && leader
-        if next_state == DPBenchmarkLifecycleState.RUNNING:
-            if len(self.peers.keys()) == 0:
-                return DPBenchmarkLifecycleState.RUNNING
-            # Now, as we are not the only unit and we are the leader, we must ensure everyone
-            # is in RUNNING state before moving forward:
-            if self.check_all_peers_in_state(DPBenchmarkLifecycleState.RUNNING):
-                return DPBenchmarkLifecycleState.RUNNING
-
-            # None of the conditions met, we return an empty state
-            return None
-
-        return next_state
 
 
 def workload_build(workload_params_template: str) -> WorkloadBase:
